Reuse rules when possible
This commit introduces a performance optimization for rules. If a rule is
marked as reusable, then it is implied that it can be instantiated only
once and reused between invocations. This should only be true for rules
which are either stateless, immutable and without side effects.

Ideally, this should be inferred. This is why we use an internal, marker
interface for this. It also makes use of a marker interface instead of
an annotation because checking for the interface is cheaper than checking
for an annotation.

Note that we could also probably avoid creating some new `InstantiatingAction`
in some cases too, but this commit seems to mitigate the performance
regression introduced by the additional attributes when improved pom support
is used.

This commit also avoids the creation of costly instantiating action rules
whenever there's no rule to execute.